+++ src/main/java/org/patryk3211/powergrid/equipment/multimeter/MultimeterItem.java
@@ -254,7 +254,7 @@ public class MultimeterItem extends Item implements IHaveElectricProperties {
                     yield 0;
                 if(!pos.isValid(level) || !neg.isValid(level))
                     yield 0;
-                float posV = 0, negV = 0;
+                double posV = 0, negV = 0;
                 if(data.contains("PosV")) {
                     posV = data.getFloat("PosV");
                 } else {
@@ -271,7 +271,7 @@ public class MultimeterItem extends Item implements IHaveElectricProperties {
                         yield 0;
                     negV = negNode.getVoltage();
                 }
-                yield posV - negV;
+                yield (float) (posV - negV);
             }
             case 1 -> {
                 var lineId = data.getInt("EID");
